implemented trait for ChatMember Privileges + fix docs and test
