+++ docs/12.features/9.dto.md
@@ -74,11 +74,13 @@ ore information on the [payment](payment#preCheckoutQuery) page.
 
 ## `ChatMember`
 
+Represents a chat member that owns the chat and has all administrator privileges. For a complete list of privileges consult https://core.telegram.org/bots/api#chatmember
+
 - `->status()` the member's status in the chat
 - `->user()` an instance of the [`User`](#user)
-- `->isAnonymous()` (optional) true, if the user's presence in the chat is hidden
-- `->isMember()` (optional) true, if the user is a member of the chat at the moment of the request
-- `->customTitle()` custom title for this user
+- `->isAnonymous()` true, if the user's presence in the chat is hidden
+- `->isMember()` true, if the user is a member of the chat at the moment of the request
+- `->customTitle()` (optional) custom title for this user
 - `->untilDate()` (optional) date when the user's subscription will expire; OR: Date when restrictions will be lifted for this user; Unix time. If 0, then the user is restricted forever
 
 ## `ChatMemberUpdate`
